fix/staking test errors

+++ programs/unloc-loan/src/constant.rs
@@ -41,6 +41,6 @@ pub const WSOL_MINT:&str = "So11111111111111111111111111111111111111112";
 pub const USDC_MINT:&str = if DEVNET_MODE {"EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v"} else {"EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v"};
 pub const UNLOC_MINT:&str = if DEVNET_MODE {"EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v"} else {"EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v"};
 #[constant]
-pub const SUB_OFFER_COUNT_PER_LEVEL: usize = 5;
+pub const SUB_OFFER_COUNT_PER_LEVEL: u64 = 5;
 #[constant]
-pub const DEFULT_SUB_OFFER_COUNT: usize = 3;
+pub const DEFULT_SUB_OFFER_COUNT: u64 = 3;
