+++ programs/unloc-loan/src/processor/process_set_sub_offer.rs
@@ -10,7 +10,7 @@ use std::str::FromStr;
 use unloc_staking::{FarmPoolUserAccount};
 pub fn process_set_sub_offer(ctx: Context<SetSubOffer>, offer_amount: u64, sub_offer_number: u64, loan_duration: u64, min_repaid_numerator: u64, apr_numerator: u64) -> Result<()> { 
     let profile_level = ctx.accounts.staking_user.profile_level;
-    let available_sub_offer_count = (DEFULT_SUB_OFFER_COUNT + profile_level * SUB_OFFER_COUNT_PER_LEVEL) as u64;
+    let available_sub_offer_count = DEFULT_SUB_OFFER_COUNT + profile_level * SUB_OFFER_COUNT_PER_LEVEL;
     require(ctx.accounts.offer.sub_offer_count - ctx.accounts.offer.start_sub_offer_num < available_sub_offer_count)?;
 
     if is_zero_account(&ctx.accounts.sub_offer.to_account_info()) {
